eval_metrics: cast bf16 → float32 before numpy in CI histograms
plot_ci_values_histograms calls .cpu().numpy() on CI tensors that arrive
as bf16 under either 3-pool (cross-pool wire dtype is bf16) or 1-pool +
autocast_bf16. NumPy doesn't support BFloat16 → TypeError on the eval
ranks, which then presents as a monitoredBarrier timeout on the other
ranks 30 min later (the actual stack trace gets buried in NCCL noise).

This was the cause of the smoke 34151 "17-min hang" attributed to a
deadlock in ~/notes/3pool_followups_2026_05_28.md task #3.

Repro'd at 8 ranks in job 34205; verified fixed in job 34207 (real fix
only, no scaffolding) running training with all five slow metrics
enabled in-train.

+++ param_decomp_lab/eval_metrics/plotting.py
@@ -469,7 +469,7 @@ def plot_ci_values_histograms(
         col = i // n_rows
         ax = axs[row, col]
 
-        data = layer_ci.flatten().cpu().numpy()
+        data = layer_ci.flatten().float().cpu().numpy()
         ax.hist(data, bins=bins)
         ax.set_yscale("log")  # Beware, memory leak unless gc.collect() is called after eval loop
         ax.set_title(f"Causal importances for {layer_name}")
